Bug 2016383 Part 2 - Make SVG private value types consistent with exported constant types r=dholbert,firefox-svg-reviewers
Differential Revision: https://phabricator.services.mozilla.com/D283009

+++ dom/svg/DOMSVGLength.cpp
@@ -361,7 +361,7 @@ void DOMSVGLength::NewValueSpecifiedUnits(uint16_t aUnit, float aValue,
       return;
     }
     AutoChangeLengthListNotifier notifier(this);
-    internalItem.SetValueAndUnit(aValue, uint8_t(aUnit));
+    internalItem.SetValueAndUnit(aValue, aUnit);
     return;
   }
   mUnit = uint8_t(aUnit);
